resolve review feedback
